pyproject.toml
[project]
name = "qc_bold"
version = "1.0.0"
description = "A reproducible pipeline for Quality Control of BOLD fMRI signals."
authors = [
    { name = "Santiago V. Blas Laguzza", email = "santiblaas@gmail.com" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21",
    "pandas>=1.3",
    "scipy>=1.7",
    "scikit-learn>=1.0",
    "matplotlib>=3.4",
    "seaborn>=0.11",
    "pyyaml>=6.0",
    "omegaconf>=2.1", # Alternativa a PyYAML para una configuración más avanzada
    "pytest>=7.0",   # Para testing
    "tqdm>=4.62",    # Para barras de progreso
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]